Split up full flow tests into different files and processed some comments
